requirements.txt
streamlit>=1.32.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
Pillow==10.1.0
scikit-image>=0.21.0
numpy>=1.24.0
opencv-python>=4.7.0
scipy>=1.10.0
